chore: remove error log for job controller tests
