+++ dhis-2/dhis-test-web-api/src/test/java/org/hisp/dhis/webapi/controller/JobConfigurationControllerTest.java
@@ -462,16 +462,6 @@ class JobConfigurationControllerTest extends H2ControllerIntegrationTestBase {
     assertStatus(HttpStatus.CONFLICT, POST("/jobConfigurations/" + jobId + "/enable"));
   }
 
-  @Test
-  void testRevert() {
-    JsonJobConfiguration config = createExpectSuccess(Map.of());
-    String jobId = config.getId();
-    switchToNewUser("no-auth");
-    assertStatus(HttpStatus.FORBIDDEN, POST("/jobConfigurations/" + jobId + "/revert"));
-    switchToAdminUser();
-    assertStatus(HttpStatus.CONFLICT, POST("/jobConfigurations/" + jobId + "/revert"));
-  }
-
   private JsonJobConfiguration createExpectSuccess(Map<String, Object> extra) {
     return createExpectSuccess(MINIMAL_CRON_CONFIG, extra);
   }
